Solution/mqtt_client/requirements.txt
aiomysql==0.0.21
aiosqlite==0.16.0
appdirs==1.4.3
astroid==2.4.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
databases==0.4.1
distlib==0.3.0
distro==1.4.0
fastapi==0.62.0
fastapi-mqtt==0.1.0
gmqtt==0.6.9
h11==0.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mqtt==0.0.1
msgpack==0.6.2
multitasking==0.0.9
numpy==1.19.4
packaging==20.3
paho-mqtt==1.5.1
pandas==1.1.5
pep517==0.8.2
progress==1.5
pydantic==1.7.3
pylint==2.6.0
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
uvicorn==0.13.1
webencodings==0.5.1
wrapt==1.12.1
yfinance==0.1.55
psycopg2-binary